change: remove emoji replacement
